+++ app/src/main/java/nextstep/signup/presentation/component/InputField.kt
@@ -1,4 +1,4 @@
-package nextstep.signup.component
+package nextstep.signup.presentation.component
 
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
