+++ app/src/main/java/nextstep/signup/presentation/theme/Color.kt
@@ -1,4 +1,4 @@
-package nextstep.signup.ui.theme
+package nextstep.signup.presentation.theme
 
 import androidx.compose.ui.graphics.Color
