+++ app/src/main/java/nextstep/signup/presentation/theme/Theme.kt
@@ -1,4 +1,4 @@
-package nextstep.signup.ui.theme
+package nextstep.signup.presentation.theme
 
 import android.os.Build
 import androidx.compose.foundation.isSystemInDarkTheme
